slightly improved depth fields, added gravity
visual artifacts are still an issue with depth fields, and the gravity can be janky. Digging to the bottom of the level can lead to odd gravity behavior and there is a small chance for the player to get their feet stuck inside the ground causing some strange collision behavior aswell.