Fixed crashing on Nvidia GPUs and untied logic from FPS.
For some reason Nvidia and Intel GPUs could encounter a problem in the frag shader where it could get stuck in an infinite loop. Adding an extra condition fixed this, but further investigation is needed to provide an optimal solution. Also, laid some groundwork for entities in the future and the game logic is no longer tied to the frame rate!